% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovb.R
\name{load_lovb_events_log}
\alias{load_lovb_events_log}
\title{Load cleaned lovb events log data from the volleydata repository.}
\usage{
load_lovb_events_log(seasons = NULL)
}
\arguments{
\item{seasons}{int, list of int, or NULL, optional
Season(s) to load. By default, NULL loads all available seasons.
- int: Single season year (e.g., 2025)
- list of int: Multiple seasons (e.g., c(2025, 2026), 2025:2026)
- NULL: Load all available seasons

\if{html}{\out{<div class="sourceCode">}}\preformatted{          All years must be 2025 or later.
}\if{html}{\out{</div>}}}
}
\value{
A data frame containing the events log data for the specified seasons.\tabular{ll}{
   Column Name \tab Type \cr
   match_id \tab int \cr
   season \tab int \cr
   match_datetime \tab string \cr
   team_name \tab string \cr
   staff_type \tab string \cr
   full_name \tab string \cr
   first_name \tab string \cr
   last_name \tab string \cr
}
}
\description{
Load cleaned lovb events log data from the volleydata repository.
}
\examples{
load_lovb_events_log(2025)
}
