% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovb.R
\name{load_lovb_officials}
\alias{load_lovb_officials}
\title{Load cleaned lovb officials data from the volleydata repository.}
\usage{
load_lovb_officials(seasons = NULL)
}
\arguments{
\item{seasons}{An integer or vector of integers of seasons to fetch data for. Defaults to all available seasons.}
}
\value{
A data frame containing the officials data for the specified seasons.\tabular{ll}{
   Column Name \tab Type \cr
   match_id \tab dbl \cr
   season \tab dbl \cr
   match_datetime \tab chr \cr
   officials_type \tab chr \cr
   full_name \tab chr \cr
   first_name \tab chr \cr
   last_name \tab chr \cr
   level \tab chr \cr
}
}
\description{
Load cleaned lovb officials data from the volleydata repository.
}
\examples{
try({load_lovb_officials(2025)})
}
