% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovb.R
\name{load_lovb_schedule}
\alias{load_lovb_schedule}
\title{Load cleaned lovb schedule data from the volleydata repository.}
\usage{
load_lovb_schedule(seasons = NULL)
}
\arguments{
\item{seasons}{An integer or vector of integers of seasons to fetch data for. Defaults to all available seasons.}
}
\value{
A data frame containing the schedule data for the specified seasons.\tabular{ll}{
   Column Name \tab Type \cr
   season \tab int \cr
   date \tab string \cr
   round \tab string \cr
   home_team \tab string \cr
   away_team \tab string \cr
   home_team_set_wins \tab int \cr
   away_team_set_wins \tab int \cr
   match_link \tab string \cr
   secondary_link \tab string \cr
   match_id \tab int \cr
}
}
\description{
Load cleaned lovb schedule data from the volleydata repository.
}
\examples{
try({load_lovb_schedule(2025)})
}
