% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlv.R
\name{load_mlv_officials}
\alias{load_mlv_officials}
\title{Load cleaned mlv officials data from the volleydata repository.}
\usage{
load_mlv_officials(seasons = NULL)
}
\arguments{
\item{seasons}{int, list of int, or NULL, optional
Season(s) to load. By default, NULL loads all available seasons.
- int: Single season year (e.g., 2025)
- list of int: Multiple seasons (e.g., c(2024, 2024), 2024:2025)
- NULL: Load all available seasons

\if{html}{\out{<div class="sourceCode">}}\preformatted{          All years must be 2024 or later.
}\if{html}{\out{</div>}}}
}
\value{
\tabular{ll}{
   Column Name \tab Type \cr
   match_id \tab int \cr
   season \tab int \cr
   match_datetime \tab string \cr
   officials_type \tab string \cr
   full_name \tab string \cr
   first_name \tab string \cr
   last_name \tab string \cr
   level \tab string \cr
}
}
\description{
Load cleaned mlv officials data from the volleydata repository.
}
\examples{
try({load_mlv_officials(2024)})

}
