% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlv.R
\name{load_mlv_pbp}
\alias{load_mlv_pbp}
\title{Load cleaned mlv points log data form the volleydata repository.}
\usage{
load_mlv_pbp(seasons = NULL)
}
\arguments{
\item{seasons}{int, list of int, or NULL, optional
Season(s) to load. By default, NULL loads all available seasons.
- int: Single season year (e.g., 2025)
- list of int: Multiple seasons (e.g., c(2024, 2024), 2024:2025)
- NULL: Load all available seasons

\if{html}{\out{<div class="sourceCode">}}\preformatted{         All years must be 2024 or later.
}\if{html}{\out{</div>}}}
}
\value{
A data frame containing the play-by-play data for the specified seasons.\tabular{ll}{
   Column Name \tab Type \cr
   match_id \tab int \cr
   season \tab int \cr
   match_datetime \tab string \cr
   home_team_name \tab string \cr
   away_team_name \tab string \cr
   team_involved \tab string \cr
   jersey_number \tab int \cr
   action \tab string \cr
   outcome \tab string \cr
   set \tab int \cr
   point_number \tab int \cr
   point_winner \tab string \cr
   home_score \tab int \cr
   away_score \tab int \cr
}
}
\description{
Load cleaned mlv points log data form the volleydata repository.
}
\examples{
load_mlv_pbp(2024)
}
