% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{load_officials}
\alias{load_officials}
\title{Load cleaned officials data from the volleydata repository.}
\usage{
load_officials(league = NULL, seasons = NULL)
}
\arguments{
\item{league}{A string specifying which of "mlv", "lovb", or "au" to load data for.}

\item{seasons}{An integer or vector of integers of seasons to fetch data for. Defaults to all available seasons.}
}
\value{
\tabular{ll}{
   Column Name \tab Type \cr
   match_id \tab int \cr
   season \tab int \cr
   match_datetime \tab string \cr
   officials_type \tab string \cr
   full_name \tab string \cr
   first_name \tab string \cr
   last_name \tab string \cr
   level \tab string \cr
}
}
\description{
Load cleaned officials data from the volleydata repository.
}
\examples{
try({load_officials("mlv", 2024)})

}
