% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlv.R
\name{load_mlv_pbp}
\alias{load_mlv_pbp}
\title{Load cleaned mlv points log data form the volleydata repository.}
\usage{
load_mlv_pbp(seasons = NULL)
}
\arguments{
\item{seasons}{An integer or vector of integers of seasons to fetch data for. Defaults to all available seasons.}
}
\value{
A data frame containing the play-by-play data for the specified seasons.\tabular{ll}{
   Column Name \tab Type \cr
   match_id \tab int \cr
   season \tab int \cr
   match_datetime \tab string \cr
   home_team_name \tab string \cr
   away_team_name \tab string \cr
   team_involved \tab string \cr
   jersey_number \tab int \cr
   action \tab string \cr
   outcome \tab string \cr
   set \tab int \cr
   point_number \tab int \cr
   point_winner \tab string \cr
   home_score \tab int \cr
   away_score \tab int \cr
}
}
\description{
Load cleaned mlv points log data form the volleydata repository.
}
\examples{
\donttest{try({load_mlv_pbp(2024)})}
}
