% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlv.R
\name{load_mlv_player_boxscore}
\alias{load_mlv_player_boxscore}
\title{Load cleaned mlv player boxscore data from the volleydata repository.}
\usage{
load_mlv_player_boxscore(seasons = NULL)
}
\arguments{
\item{seasons}{An integer or vector of integers of seasons to fetch data for. Defaults to all available seasons.}
}
\value{
A data frame containing player boxscore data for the specified seasons.
|Column Name              |  Type |
|match_id                 |  int  |
|season                   |  int  |
|match_datetime           |  chr  |
|team_involved            |  chr  |
|team_name                |  chr  |
|player_name              |  chr  |
|last_name                |  chr  |
|first_name               |  chr  |
|sets_played              |  int  |
|player_number            |  chr  |
|is_captain               |  lgl  |
|is_libero                |  int  |
|set_starting_position    |  chr  |
|set_number               |  int  |
|serves                   |  int  |
|serve_errors             |  int  |
|serve_aces               |  int  |
|serve_efficiency         |  dbl  |
|attack_attempts          |  int  |
|attack_errors            |  int  |
|attack_kills             |  int  |
|attack_success_ratio     |  dbl  |
|attack_efficiency        |  dbl  |
|receptions               |  int  |
|reception_errors         |  int  |
|positive_reception_ratio |  dbl  |
|perfect_reception_ratio  |  dbl  |
|block_points             |  int  |
|block_touches            |  int  |
|earned_points            |  int  |
|net_points               |  int  |
|assists                  |  int  |
|successful_digs          |  int  |
|id                       |  chr  |
|spike_hp                 |  int  |
|points                   |  int  |
}
\description{
Load cleaned mlv player boxscore data from the volleydata repository.
}
\examples{
try({load_mlv_player_boxscore(2024)})
}
