% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlv.R
\name{load_mlv_schedule}
\alias{load_mlv_schedule}
\title{Load cleaned mlv schedule data from the volleydata repository.}
\usage{
load_mlv_schedule(seasons = NULL)
}
\arguments{
\item{seasons}{An integer or vector of integers of seasons to fetch data for. Defaults to all available seasons.

\if{html}{\out{<div class="sourceCode">}}\preformatted{           All years must be 2024 or later.
}\if{html}{\out{</div>}}}
}
\value{
\tabular{ll}{
   Column Name \tab Type \cr
   season \tab int \cr
   date \tab string \cr
   round \tab string \cr
   home_team \tab string \cr
   away_team \tab string \cr
   home_team_set_wins \tab int \cr
   away_team_set_wins \tab int \cr
   result \tab string \cr
   match_link \tab string \cr
}
}
\description{
Load cleaned mlv schedule data from the volleydata repository.
}
\examples{
try({load_mlv_schedule(2024)})
}
