% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlv.R
\name{load_mlv_team_staff}
\alias{load_mlv_team_staff}
\title{Load cleaned mlv team staff data from the volleydata repository.}
\usage{
load_mlv_team_staff(seasons = NULL)
}
\arguments{
\item{seasons}{int, list of int, or NULL, optional
Season(s) to load. By default, NULL loads all available seasons.
- int: Single season year (e.g., 2025)
- list of int: Multiple seasons (e.g., c(2024, 2024), 2024:2025)
- NULL: Load all available seasons

\if{html}{\out{<div class="sourceCode">}}\preformatted{         All years must be 2024 or later.
}\if{html}{\out{</div>}}}
}
\value{
A data frame containing the team staff data for the specified seasons.\tabular{ll}{
   Column Name \tab Type \cr
   match_id \tab dbl \cr
   season \tab dbl \cr
   match_datetime \tab chr \cr
   team_name \tab chr \cr
   staff_type \tab chr \cr
   full_name \tab chr \cr
   first_name \tab chr \cr
   last_name \tab chr \cr
}
}
\description{
Load cleaned mlv team staff data from the volleydata repository.
}
\examples{
load_mlv_team_staff(2024)
}
